refactor(climate): Extract climate entities into dedicated modules
Move PanasonicClimateEntity and AquareaClimateEntity classes from
climate.py into new entities/panasonic_climate.py and
entities/aquarea_climate.py respectively.

- Create entities/ subpackage with separate files for each climate entity type
- Reduce climate.py to a thin setup module (~30 lines) that imports and
  instantiates entities from the new modules
- Relocate all entity classes, helper functions, constants, and
  entity descriptions to their respective new files
- Maintain identical behavior — pure code reorganization with no
  functional changes
