+++ custom_components/panasonic_cc/entities/__init__.py
@@ -1,0 +1,1 @@
+"""Entity submodules for Panasonic Comfort Cloud."""
